move InputFilter components to elements
